refactor: mui theme using
