+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ats-react-app",
+  "name": "uc-ats-react-app",
   "private": true,
   "version": "0.0.0",
   "type": "module",
