--- src/components/layouts/index.ts
@@ -1,3 +1,0 @@
-export { default as BaseLayout } from "./BaseLayout"
-export { default as AsideNavigateLayout } from "./AsideNavigateLayout"
-export { default as JobDetailLayout } from "./JobDetailLayout"
