+++ src/components/sections/TheNavigateAsideBar.tsx
@@ -2,14 +2,12 @@ import Stack from "@mui/material/Stack"
 import Box from "@mui/material/Box"
 import { useMemo } from "react"
 import { useLocation, useNavigate } from "react-router-dom"
-import { useTheme } from "@mui/material"
 import { APP_HEADER_HEIGHT } from "constants/STYLE"
 import { IconJob, IconPeople, IconStoreBox } from "assets/icons"
 
 const TheNavigateAsideBar = () => {
   const navigate = useNavigate()
   const { pathname } = useLocation()
-  const { palette } = useTheme()
 
   const navigateList = useMemo(
     () => [
@@ -36,9 +34,7 @@ const TheNavigateAsideBar = () => {
   )
 
   return (
-    <Box
-      sx={{ minWidth: 304, borderRight: "1px solid", borderRightColor: palette.border.outlined }}
-    >
+    <Box sx={{ minWidth: 304, borderRight: "1px solid", borderRightColor: "border.outlined" }}>
       <Stack sx={{ position: "sticky", top: APP_HEADER_HEIGHT, p: "24px", gap: "12px" }}>
         {navigateList.map((item) => (
           <Stack
@@ -51,13 +47,13 @@ const TheNavigateAsideBar = () => {
               borderRadius: "12px",
               px: "16px",
               cursor: "pointer",
-              color: item.isActive && palette.text.accent,
-              bgcolor: item.isActive && palette.background.accent,
+              color: item.isActive && "text.accent",
+              bgcolor: item.isActive && "background.accent",
               filter: !item.isActive && "brightness(0) contrast(100%)",
               duration: 0.2,
               "&:hover": {
-                bgcolor: palette.background.accent,
-                color: palette.text.accent,
+                bgcolor: "background.accent",
+                color: "text.accent",
                 filter: "none",
               },
             }}
