+++ src/components/sections/home/TopBarWithTitle.tsx
@@ -1,6 +1,5 @@
 import Stack from "@mui/material/Stack"
 import Typography from "@mui/material/Typography"
-import { useTheme } from "@mui/material"
 import { ButtonContained } from "components/common"
 import { APP_HEADER_HEIGHT } from "constants/STYLE"
 import { UCT_LINK_PAGE } from "constants/UCT_WEB"
@@ -8,7 +7,6 @@ import { openLinkInNewTab } from "utils/common"
 import { IconPlus } from "assets/icons"
 
 const TopBarWithTitle = () => {
-  const { palette } = useTheme()
   const onClickPostJob = () => {
     openLinkInNewTab(UCT_LINK_PAGE.POST_JOB)
   }
@@ -22,7 +20,7 @@ const TopBarWithTitle = () => {
         position: "sticky",
         top: APP_HEADER_HEIGHT,
         px: "28px",
-        bgcolor: palette.background.accent,
+        bgcolor: "background.accent",
         zIndex: 10,
       }}
     >
