+++ src/layouts/AsideNavigateLayout.tsx
@@ -1,7 +1,6 @@
 import Stack from "@mui/material/Stack"
 import Box from "@mui/material/Box"
 import useScrollToTop from "hooks/useScrollToTop"
-import { useTheme } from "@mui/material"
 import { AppContainer } from "components/common"
 import { TheAppHeader, TheNavigateAsideBar } from "components/sections"
 import { type ReactNode } from "react"
@@ -11,7 +10,6 @@ interface TypeProps {
 }
 
 const AsideNavigateLayout: React.FC<TypeProps> = ({ children }) => {
-  const { palette } = useTheme()
   const { containerRef } = useScrollToTop()
 
   return (
@@ -25,7 +23,7 @@ const AsideNavigateLayout: React.FC<TypeProps> = ({ children }) => {
               minHeight: "100%",
               borderWidth: "0 1px",
               borderStyle: "solid",
-              borderColor: palette.border.outlined,
+              borderColor: "border.outlined",
             }}
           >
             <TheNavigateAsideBar />
