+++ src/pages/TheHomePage.tsx
@@ -1,10 +1,7 @@
 import Box from "@mui/material/Box"
-import { useTheme } from "@mui/material"
 import { TopBarWithTitle, TheJobsList } from "components/sections/home"
 
 const TheHomePage = () => {
-  const { palette } = useTheme()
-
   return (
     <>
       <TopBarWithTitle />
@@ -15,7 +12,7 @@ const TheHomePage = () => {
           px: "28px",
           position: "sticky",
           bottom: 0,
-          bgcolor: palette.background.accent,
+          bgcolor: "background.accent",
         }}
       ></Box>
     </>
